Update app.py
Added a single line of code to prevent Flask from caching files. Without this, modifying CS/JS files, pushing to flip, and restarting the Gunicorn server *will not* serve the updated files. This is because Flask has cached them, and restarting Gunicorn does not restart Flask.

Gunicorn was not part of the Flask demo in this tutorial, but was presented as a solution by Danielle Safonte on Piazza because Flask does not run correctly on the flip (unix) servers. I confirmed that when my own Flask instance didn't run correctly, and implemented Gunicorn as a WSGI in front of Flask.

Putting this line of code in resolved my caching issues. Gunicorn still needs to be restarted to display updated CSS/JS files. This is specific to Flask and will work with or without Gunicorn.